feat(question-type-cursor): #FOR-498 review
