+++ formulaire/src/main/java/fr/openent/formulaire/service/QuestionSpecificFieldService.java
@@ -7,6 +7,13 @@ import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 
 public interface QuestionSpecificFieldService {
+
+    /**
+     * Sync all the specifics fields of question
+     * @param questions JsonArray data
+     */
+    Future<JsonArray> syncQuestionSpecs(JsonArray questions);
+
     /**
      * List all the specifics fields of question from a list of ids
      * @param questionIds questions identifiers
