+++ formulaire/src/main/java/fr/openent/formulaire/service/impl/DefaultQuestionSpecificFieldService.java
@@ -3,6 +3,8 @@ package fr.openent.formulaire.service.impl;
 import fr.openent.form.core.constants.Fields;
 import fr.openent.form.core.enums.QuestionTypes;
 import fr.openent.form.helpers.FutureHelper;
+import fr.openent.form.helpers.UtilsHelper;
+import fr.openent.formulaire.controllers.QuestionController;
 import fr.openent.formulaire.service.QuestionSpecificFieldService;
 import fr.wseduc.webutils.Either;
 import io.vertx.core.Future;
@@ -10,15 +12,39 @@ import io.vertx.core.Handler;
 import io.vertx.core.Promise;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
+import io.vertx.core.logging.Logger;
+import io.vertx.core.logging.LoggerFactory;
 import org.entcore.common.sql.Sql;
 import org.entcore.common.sql.SqlResult;
 import org.entcore.common.sql.SqlStatementsBuilder;
 
 import static fr.openent.form.core.constants.Fields.*;
 import static fr.openent.form.core.constants.Tables.*;
+import static fr.openent.form.helpers.UtilsHelper.getIds;
 
 public class DefaultQuestionSpecificFieldService implements QuestionSpecificFieldService {
     private final Sql sql = Sql.getInstance();
+
+    private static final Logger log = LoggerFactory.getLogger(QuestionController.class);
+
+    public Future<JsonArray> syncQuestionSpecs(JsonArray questions) {
+        Promise<JsonArray> promise = Promise.promise();
+
+        JsonArray questionIds = getIds(questions);
+        if (!questions.isEmpty()) {
+            listByIds(questionIds)
+                    .onSuccess(specEvt -> {
+                        promise.complete(UtilsHelper.mergeQuestionsAndSpecifics(questions, specEvt));
+                    })
+                    .onFailure(error -> {
+                        String message = String.format("[Formulaire@%s::syncQuestionSpecs] An error has occured" +
+                                " when getting specific field: %s", this.getClass().getSimpleName(), error.getMessage());
+                        log.error(message, error.getMessage());
+                        promise.fail(error.getMessage());
+                    });
+        } else promise.complete(new JsonArray());
+        return promise.future();
+    }
 
     @Override
     public Future<JsonArray> listByIds(JsonArray questionIds) {
