+++ formulaire/src/test/java/fr/openent/formulaire/service/test/impl/DefaultQuestionServiceTest.java
@@ -80,7 +80,10 @@ public class DefaultQuestionServiceTest {
             ctx.assertEquals(expectedParams.toString(), body.getJsonArray(VALUES).toString());
             async.complete();
         });
-        defaultQuestionService.listForFormAndSection(FORM_ID);
+        defaultQuestionService.listForFormAndSection(FORM_ID)
+                .onSuccess(result -> async.complete());
+
+        async.awaitSuccess(10000);
     }
 
     @Test
